chore

+++ contracts/Timelock.sol
@@ -10,21 +10,34 @@ contract TimeLock {
 
     struct LockBoxStruct {
         address beneficiary;
-        uint balance;
-        uint releaseTime;
+        uint256 balance;
+        uint256 releaseTime;
     }
 
     LockBoxStruct[] public lockBoxStructs; // This could be a mapping by address, but these numbered lockBoxes support possibility of multiple tranches per address
 
-    event LogLockBoxDeposit(address sender, uint amount, uint releaseTime);   
-    event LogLockBoxTransfer(address sender, address to, uint amount, uint releaseTime);   
-    event LogLockBoxWithdrawal(address receiver, uint amount);
+    event LogLockBoxDeposit(
+        address sender,
+        uint256 amount,
+        uint256 releaseTime
+    );
+    event LogLockBoxTransfer(
+        address sender,
+        address to,
+        uint256 amount,
+        uint256 releaseTime
+    );
+    event LogLockBoxWithdrawal(address receiver, uint256 amount);
 
     constructor(address tokenContract) public {
         token = IERC20(tokenContract);
     }
 
-    function deposit(address beneficiary, uint amount, uint releaseTime) public {
+    function deposit(
+        address beneficiary,
+        uint256 amount,
+        uint256 releaseTime
+    ) public {
         require(token.transferFrom(msg.sender, address(this), amount));
         LockBoxStruct memory l;
         l.beneficiary = beneficiary;
@@ -32,24 +45,28 @@ contract TimeLock {
         l.releaseTime = releaseTime;
         lockBoxStructs.push(l);
         emit LogLockBoxDeposit(msg.sender, amount, releaseTime);
-
     }
 
-    function withdraw(uint lockBoxNumber) public {
+    function withdraw(uint256 lockBoxNumber) public {
         LockBoxStruct storage l = lockBoxStructs[lockBoxNumber];
         require(l.beneficiary == msg.sender);
         require(l.releaseTime <= block.timestamp);
-        uint amount = l.balance;
+        uint256 amount = l.balance;
         l.balance = 0;
         emit LogLockBoxWithdrawal(msg.sender, amount);
         require(token.transfer(msg.sender, amount));
     }
 
-    function transferLockedFunds(uint lockBoxNumber, address beneficiary, uint amount, uint releaseTime) public {
+    function transferLockedFunds(
+        uint256 lockBoxNumber,
+        address beneficiary,
+        uint256 amount,
+        uint256 releaseTime
+    ) public {
         LockBoxStruct storage l = lockBoxStructs[lockBoxNumber];
         require(l.beneficiary == msg.sender);
         require(releaseTime >= l.releaseTime);
-        uint256 balance =  l.balance;
+        uint256 balance = l.balance;
         l.balance = balance.sub(amount);
 
         LockBoxStruct memory newL;
@@ -58,6 +75,6 @@ contract TimeLock {
         newL.releaseTime = releaseTime;
         lockBoxStructs.push(newL);
 
-        emit LogLockBoxTransfer(msg.sender, beneficiary,amount, releaseTime);
+        emit LogLockBoxTransfer(msg.sender, beneficiary, amount, releaseTime);
     }
 }
